// dao-governance/stack/dao/app/src/types/roles.tsx
export interface DAORoleHat extends Omit<DAOHat, 'smartAddress'> {
  wearerAddress: Address;
  smartAddress?: Address;
  roleTerms: DAORoleHatTerms;
  canCreateProposals: boolean;
  payments: SablierPayment[];
  isTermed: boolean;
  eligibility?: Address;
}
// ...
export interface SablierPaymentFormValues extends Partial<SablierPayment> {
  isStreaming: () => boolean;
  canUserCancel: () => boolean;
  isCancelling: boolean;
  isValidatedAndSaved?: boolean;
  cancelable: boolean;
}
// ...
export interface EditedRole {
  fieldNames: EditedRoleFieldNames[];
  status: EditBadgeStatus;
}

export interface RoleHatFormValue
  extends Partial<Omit<DAORoleHat, 'id' | 'wearerAddress' | 'payments' | 'roleTerms'>> {
  id: Hex;
  // The user-input field that could either be an address or an ENS name.
  wearer?: string;
  // Not a user-input field.
  // `resolvedWearer` is dynamically set from the resolved address of `wearer`, in case it's an ENS name.
  resolvedWearer?: Address;
  payments: SablierPaymentFormValues[];
  // form specific state
  editedRole?: EditedRole;
  roleEditingPaymentIndex?: number;
  isTermed?: boolean;
  roleTerms?: {
    nominee?: string;
    termEndDate?: Date;
    termNumber: number;
  }[];
  canCreateProposals: boolean;
}

export interface RoleHatFormValueEdited extends RoleHatFormValue {
  editedRole: EditedRole;
}

export type RoleFormValues = {
  proposalMetadata: CreateProposalMetadata;
  hats: RoleHatFormValue[];
  roleEditing?: RoleHatFormValue;
  customNonce?: number;
  actions: SendAssetsData[];
  newRoleTerm?: {
    nominee: string;
    termEndDate: Date;
    termNumber: number;
  };
};